Various Ember upgrade fixes.
