+++ src/api-umbrella/admin-ui/app/components/apis/sub-settings-form.js
@@ -12,22 +12,18 @@ import classic from 'ember-classic-decorator';
 export default class SubSettingsForm extends Component {
   openModal = false;
 
-  init() {
-    super.init(...arguments);
-
-    this.httpMethodOptions = [
-      { id: 'any', name: 'Any' },
-      { id: 'GET', name: 'GET' },
-      { id: 'POST', name: 'POST' },
-      { id: 'PUT', name: 'PUT' },
-      { id: 'DELETE', name: 'DELETE' },
-      { id: 'HEAD', name: 'HEAD' },
-      { id: 'TRACE', name: 'TRACE' },
-      { id: 'OPTIONS', name: 'OPTIONS' },
-      { id: 'CONNECT', name: 'CONNECT' },
-      { id: 'PATCH', name: 'PATCH' },
-    ];
-  }
+  httpMethodOptions = [
+    { id: 'any', name: 'Any' },
+    { id: 'GET', name: 'GET' },
+    { id: 'POST', name: 'POST' },
+    { id: 'PUT', name: 'PUT' },
+    { id: 'DELETE', name: 'DELETE' },
+    { id: 'HEAD', name: 'HEAD' },
+    { id: 'TRACE', name: 'TRACE' },
+    { id: 'OPTIONS', name: 'OPTIONS' },
+    { id: 'CONNECT', name: 'CONNECT' },
+    { id: 'PATCH', name: 'PATCH' },
+  ];
 
   @computed('model.isNew')
   get modalTitle() {
