--- src/api-umbrella/admin-ui/app/mixins/sortable.js
@@ -1,54 +1,0 @@
-import { computed } from '@ember/object';
-import { guidFor } from '@ember/object/internals';
-import Mixin from '@ember/object/mixin'
-import { t } from 'api-umbrella-admin-ui/utils/i18n';
-import sortable from 'html5sortable/dist/html5sortable.es';
-
-// eslint-disable-next-line ember/no-new-mixins
-export default Mixin.create({
-  isReorderable: computed('sortableCollection.[]', function() {
-    const length = this.sortableCollection.length;
-    return (length > 1);
-  }),
-
-  updateSortOrder(indexes) {
-    this.sortableCollection.forEach(function(record) {
-      const index = indexes[guidFor(record)];
-      record.set('sortOrder', index + 1);
-    });
-  },
-
-  actions: {
-    reorderCollection(containerId) {
-      const container = document.getElementById(containerId);
-      const buttonText = container.querySelector('.reorder-button-text');
-
-      if(container.classList.contains('reorder-active')) {
-        buttonText.innerText = buttonText.dataset.originalText;
-        container.classList.remove('reorder-active');
-      } else {
-        buttonText.dataset.originalText = buttonText.innerText;
-        buttonText.innerText = t('Done');
-        container.classList.add('reorder-active');
-      }
-
-      const tbody = container.querySelector('tbody');
-      sortable(tbody, {
-        items: 'tr',
-        handle: '.reorder-handle',
-        forcePlaceholderSize: true,
-        placeholderClass: 'reorder-placeholder',
-      });
-      tbody.addEventListener('sortupdate', () => {
-        const indexes = {};
-        const rows = tbody.querySelectorAll('tr');
-        for(let i = 0; i < rows.length; i++) {
-          const row = rows[i];
-          indexes[row.dataset.guid] = i;
-        }
-
-        this.updateSortOrder(indexes);
-      });
-    },
-  },
-});
