+++ src/api-umbrella/admin-ui/app/templates/components/apis/sub-settings-table.hbs
@@ -33,8 +33,8 @@
       <button type="button" class="btn btn-light btn-sm" {{action "add"}}><FaIcon @icon="plus-circle" /> Add URL Settings</button>
     </div>
     <div class="col-6 text-right">
-      {{#if this.isReorderable}}
-        <button type="button" id="sub_settings_reorder" class="btn btn-light btn-sm" {{action "reorderCollection" "sub_settings"}}><FaIcon @icon="bars" /> <span class="reorder-button-text">Reorder</span></button>
+      {{#if this.sortable.isReorderable}}
+        <button type="button" id="sub_settings_reorder" class="btn btn-light btn-sm" {{on "click" (fn this.sortable.reorderCollection "sub_settings")}}><FaIcon @icon="bars" /> <span class="reorder-button-text">Reorder</span></button>
       {{/if}}
     </div>
   </div>
